refactor(tests): split useVersionCheck.test.tsx into 3 focused files
Splits the 2167-line monolith into:
- useVersionCheck-parse-provider: parseReleaseTag, parseRelease, getLatestForChannel, cache behaviour, VersionCheckProvider, toggle-sensitive polling
- useVersionCheck-version-compare: isNewerVersion (via hasUpdate), version comparison edge cases, isDevVersion, isNewerVersion, skipVersion, setChannel, triggerUpdate, fetchReleases edge cases
- useVersionCheck-fetch-update: fetchLatestMainSHA, forceCheck, fetchAutoUpdateStatus, loadChannel, loadCache, installMethod, setAutoUpdateEnabled, checkForUpdates, backend install method detection, helm install, fetchRecentCommits

Fixes #8624
